brea 机器升级模块系统正在开发

+++ ArkdustEndField/src/main/java/com/landis/arkdust/blockentity/thermo/ThermoCarrierBlockEntity.java
@@ -12,7 +12,7 @@ public class ThermoCarrierBlockEntity extends ThermoBlockEntity {
     }
 
     @Override
-    public int getM() {
+    public long getM() {
         return 1000;
     }
